chore: 블로르 댓글 컴포넌트 분리
