+++ src/pages/Blog/BlogDetail.jsx
@@ -2,18 +2,7 @@ import { useState, useEffect } from 'react';
 import { useParams, useLocation } from 'react-router-dom';
 import styled from 'styled-components';
 import GlobalStyle from '@/styles/global';
-import {
-  Button,
-  Image,
-  Toast,
-  Header,
-  dummyData,
-  Modal,
-  ModalText,
-  ButtonContainer,
-} from '@/components';
-import { useLogin } from '@/context/LoginContext';
-import { MoreIcon } from '@/assets';
+import { Image, Toast, Header, dummyData, BlogComment } from '@/components';
 
 const DetailContainer = styled.div`
   display: flex;
@@ -68,70 +57,6 @@ const Content = styled.div`
   word-break: keep-all;
 `;
 
-const CommentContent = styled.div`
-  max-width: 720px;
-  width: 100%;
-  margin-top: 25px;
-`;
-
-const CommentHeader = styled.div`
-  font-size: 16px;
-  font-weight: 500;
-  margin-bottom: 16px;
-
-  span {
-    color: #007aff;
-    margin-left: 4px;
-  }
-`;
-
-const NoComment = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  font-size: 14px;
-  line-height: 0;
-  color: #c8c8c8;
-  margin: 50px 0;
-`;
-
-const CommentInput = styled.textarea`
-  width: 100%;
-  height: 80px;
-  padding: 25px 16px;
-  font-size: 14px;
-  color: #333333;
-  background-color: #ffffff;
-  border: 1px solid #e6e6e6;
-  border-radius: 4px;
-  resize: none;
-  outline: none;
-
-  &::placeholder {
-    color: #909090;
-  }
-`;
-
-const LoginCommentInput = styled.textarea`
-  width: 100%;
-  height: 100px;
-  padding-top: 20px;
-  margin-top: 10px 0 -10px;
-  font-size: 14px;
-  color: #333333;
-  background-color: #ffffff;
-  border: none;
-  border-bottom: 1px solid #e6e6e6;
-  border-radius: 4px;
-  resize: none;
-  outline: none;
-  word-break: break-word;
-
-  &::placeholder {
-    color: #909090;
-  }
-`;
-
 const Nickname = styled.div`
   padding-top: 20px;
   font-size: 24px;
@@ -164,59 +89,7 @@ const ProfileContent = styled.div`
   width: 720px;
 `;
 
-const CommentBox = styled.div`
-  display: flex;
-  flex-direction: column;
-  border: 1px solid #e6e6e6;
-  border-radius: 4px;
-  padding: 16px;
-`;
-
-const PostedCommetBox = styled.div`
-  display: flex;
-  flex-direction: column;
-  margin: 50px 0 60px;
-`;
-
-const PostedCommentContent = styled.div`
-  display: flex;
-  justify-content: space-between;
-`;
-
-const CommentData = styled.p`
-  display: flex;
-  flex-direction: column;
-  font-size: 12px;
-  color: #909090;
-  padding-left: 27px;
-  margin-top: 0px;
-`;
-
-const PostedComment = styled.p`
-  font-size: 14px;
-  font-weight: 300;
-  color: #333;
-  line-height: 160%;
-  letter-spacing: -0.5%;
-  padding-left: 25px;
-  overflow-wrap: break-word;
-  word-break: break-all;
-  white-space: pre-wrap;
-`;
-
-const ButtonWrapper = styled.div`
-  display: flex;
-  justify-content: flex-end;
-`;
-
-const StyledMoreIcon = styled(MoreIcon)`
-  &:hover {
-    opacity: 0.6;
-  }
-`;
-
 const BlogDetail = () => {
-  const { isLogin } = useLogin();
   const location = useLocation();
 
   const { id } = useParams();
@@ -234,39 +107,6 @@ const BlogDetail = () => {
       return () => clearTimeout(timer);
     }
   }, [toastData]);
-
-  const [commentCount, setCommentCount] = useState(post?.comments || 0);
-  const [commentList, setCommentList] = useState([]);
-  const [commentInput, setCommentInput] = useState('');
-
-  const PostComment = () => {
-    if (!commentInput.trim()) return;
-    const newComment = {
-      id: Date.now(), // 유니크한 값을 id로
-      text: commentInput,
-    };
-    setCommentList((prev) => [...prev, newComment]);
-    setCommentCount((prev) => prev + 1);
-    setCommentInput('');
-  };
-
-  const [modalOpen, setModalOpen] = useState(false);
-  const [deleteId, setDeleteId] = useState(null);
-
-  const deleteComment = () => {
-    setCommentList((prev) => prev.filter((comment) => comment.id !== deleteId));
-    setCommentCount((prev) => prev - 1);
-    setModalOpen(false);
-    setDeleteId(null);
-  };
-
-  const onToast = () => {
-    setToastData({
-      show: true,
-      type: 'positive',
-      message: '삭제가 완료되었습니다!',
-    });
-  };
 
   if (!post) {
     return <p>게시글을 찾을 수 없습니다.</p>;
@@ -291,62 +131,7 @@ const BlogDetail = () => {
         <Line />
         <Content>{post.content}</Content>
         <Line />
-        <CommentContent id='comment'>
-          <CommentHeader>
-            댓글 <span>{commentList.length}</span>
-          </CommentHeader>
-          {commentCount === 0 && (
-            <NoComment>
-              <p>작성된 댓글이 없습니다.</p>
-              <p>응원의 첫 번째 댓글을 달아주세요.</p>
-            </NoComment>
-          )}
-          {commentList.map((comment) => (
-            <PostedCommetBox key={comment.id}>
-              <PostedCommentContent>
-                <InfoWrapper>
-                  <Image width='20px' height='20px' src={post.profileImg} alt='프로필' />
-                  <NameText>{post.nickname}</NameText>
-                </InfoWrapper>
-                <StyledMoreIcon
-                  onClick={() => {
-                    setDeleteId(comment.id);
-                    setModalOpen(true);
-                  }}
-                />
-              </PostedCommentContent>
-              <CommentData>{post.date}</CommentData>
-              <PostedComment>{comment.text}</PostedComment>
-            </PostedCommetBox>
-          ))}
-          {isLogin ? (
-            <CommentBox>
-              <InfoWrapper>
-                <Image width='20px' height='20px' src={post.profileImg} alt='프로필' />
-                <NameText>{post.nickname}</NameText>
-              </InfoWrapper>
-              <LoginCommentInput
-                placeholder='댓글을 입력해주세요.'
-                value={commentInput}
-                onChange={(e) => setCommentInput(e.target.value)}
-              />
-              <br />
-              <ButtonWrapper>
-                <Button
-                  width='65px'
-                  borderStyle='1px solid #909090'
-                  color='#909090'
-                  radius='25px'
-                  onClick={PostComment}
-                >
-                  등록
-                </Button>
-              </ButtonWrapper>
-            </CommentBox>
-          ) : (
-            <CommentInput disabled placeholder='로그인을 하고 댓글을 달아보세요!' />
-          )}
-        </CommentContent>
+        <BlogComment post={post} />
         <ProfileBox>
           <ProfileContent>
             <Image src={post.profileImg} alt='프로필' width='64px' height='64px' radius='50%' />
@@ -355,34 +140,6 @@ const BlogDetail = () => {
           </ProfileContent>
         </ProfileBox>
       </DetailContainer>
-      <Modal isOpen={modalOpen}>
-        <ModalText>
-          <h4>댓글을 삭제할까요?</h4>
-        </ModalText>
-        <ButtonContainer>
-          <Button
-            onClick={() => setModalOpen(false)}
-            width='150px'
-            borderStyle='1px solid #dfdada'
-            radius='3px'
-          >
-            취소
-          </Button>
-          <Button
-            width='150px'
-            borderStyle='none'
-            radius='3px'
-            color='white'
-            bgColor='#FF3F3F'
-            onClick={() => {
-              deleteComment();
-              onToast();
-            }}
-          >
-            삭제하기
-          </Button>
-        </ButtonContainer>
-      </Modal>
     </>
   );
 };
